cast request_start to float instead of integer

+++ lib/rack/timeout.rb
@@ -25,7 +25,7 @@ module Rack
       info          = env[ENV_INFO_KEY] ||= RequestDetails.new
       info.id     ||= env['HTTP_HEROKU_REQUEST_ID'] || env['HTTP_X_REQUEST_ID'] || SecureRandom.hex
       request_start = env['HTTP_X_REQUEST_START'] # unix timestamp in ms
-      request_start = Time.at(request_start.to_i / 1000) if request_start
+      request_start = Time.at(request_start.to_f / 1000) if request_start
       info.age      = Time.now - request_start           if request_start
       has_body      = env["rack.input"].size > 0         if env["rack.input"]
       time_left     = MAX_REQUEST_AGE - info.age         if info.age
